Syncing with core: Minor adjustment to TMUnit drop migration

+++ pootle/apps/pootle_store/migrations/0015_auto__del_tmunit.py
@@ -2,14 +2,17 @@
 from south.utils import datetime_utils as datetime
 from south.db import db
 from south.v2 import SchemaMigration
-from django.db import models
+from django.db import connection, models
+from django.contrib.contenttypes.models import ContentType
 
 
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
-        # Deleting model 'TMUnit'
-        db.delete_table(u'pootle_store_tmunit')
+        if u'pootle_store_tmunit' in connection.introspection.table_names():
+            # Deleting model 'TMUnit'
+            ContentType.objects.filter(app_label='pootle_store', model='tmunit').delete()
+            db.delete_table(u'pootle_store_tmunit')
 
 
     def backwards(self, orm):
